# central-server/app/routes.py
from fastapi import APIRouter, HTTPException
from models import Job, Worker, JobSubmission, JobStatus, WorkerHeartbeat

router = APIRouter()

@router.post("/jobs", response_model=Job)
async def submit_job(job_submission: JobSubmission):
    """
    Submit a new job to the system.
    """
    # Pseudo code for job submission
    return {"id": "job_id", **job_submission.dict()}

@router.get("/jobs/{job_id}", response_model=JobStatus)
async def get_job_status(job_id: str):
    """
    Retrieve the status of a job.
    """
    # Pseudo code for retrieving job status
    return {"job_id": job_id, "status": "JobStatus"}

@router.get("/workers", response_model=list[Worker])
async def list_workers():
    """
    List all registered workers.
    """
    # Pseudo code for listing workers
    return [{"id": "worker_id", "ip_address": "192.168.1.1", "capacity": 10, "utilization": 5}]

@router.post("/workers/{worker_id}/heartbeat", response_model=WorkerHeartbeat)
async def worker_heartbeat(worker_id: str, heartbeat: WorkerHeartbeat):
    """
    Receive a heartbeat from a worker.
    """
    # Pseudo code for handling worker heartbeat
    return {"worker_id": worker_id, **heartbeat.dict()}